Create One million headlines in one minute
The most common words from 1M ABC news headlines visualised as a word map

More frequently used words for each month are in darker, larger text. The location of the words is based on a longer time frame (6 months) for a smoother visual, with the more common words towards the centre. The code takes ~85 seconds to run on my computer. The bulk of this is the groupby().sum() function on line 26 - it takes around 76 seconds. 

The data set is from https://www.kaggle.com/therohk/million-headlines and contains 1,226,258 ABC news headlines. 
The stop words are based on the default set at https://www.ranks.nl/stopwords.